fixed crashes on second levels, need to think about not using long long for all of the sun variables but it works for now.

+++ orbits.cpp
@@ -25,7 +25,7 @@ Body::Body(double xPosition, double yPosition, double bodyMass, double bodyRadiu
 
 //Sun constructor
 //-----------------------------------------------------------------
-Sun::Sun(double xPosition, double yPosition, double bodyMass, double bodyRadius, int colorRed, int colorGreen, int colorBlue) : Body( xPosition,  yPosition, bodyMass, bodyRadius, colorRed, colorGreen, colorBlue){
+Sun::Sun(double xPosition, double yPosition, long long bodyMass, double bodyRadius, int colorRed, int colorGreen, int colorBlue) : Body( xPosition,  yPosition, bodyMass, bodyRadius, colorRed, colorGreen, colorBlue){
   
 }
